Changed srcDirs to sourceFiles.
