+++ demo/Calculator/builds/linux-clang++-debug/target.toml
@@ -18,14 +18,14 @@ flags = [
     name = "calculatorstatic"
     buildRule = "staticLib"
     outputName = "CalculatorStatic"
-    srcDirs = ["lib"]
+    sourceFiles = ["lib/*.cpp"]
     includePaths = ["include"]
 
 [[builds]]
     name = "calculatordynamic"
     buildRule = "dynamicLib"
     outputName = "CalculatorShared"
-    srcDirs = ["lib"]
+    sourceFiles = ["lib/*.cpp"]
     includePaths = ["include"]
 
 [[builds]]
@@ -33,7 +33,7 @@ flags = [
     buildRule = "exe"
     requires = ["calculatorstatic"]
     outputName = "CalculatorTests"
-    srcDirs = ["tests"]
+    sourceFiles = ["tests/*.cpp"]
     includePaths = ["include"]
     # The libraryPaths and libraries fields can be used to specify additional
     # libraries and paths to the build. This allows for linking against third
@@ -46,5 +46,5 @@ flags = [
     buildRule = "exe"
     requires = ["calculatordynamic"]
     outputName = "CalculatorApp"
-    srcDirs = ["src"]
+    sourceFiles = ["src/*.cpp"]
     includePaths = ["include"]
