+++ demo/Empty/builds/windows-clang++-debug/target.toml
@@ -12,7 +12,7 @@ flags = [
     name = "libraryname"
     buildRule = "staticLib" # or dynamicLib
     outputName = "LibraryName"
-    srcDirs = ["lib"]
+    sourceFiles = ["lib/*.cpp"]
     includePaths = ["include"]
 
 [[builds]]
@@ -20,7 +20,7 @@ flags = [
     buildRule = "exe"
     requires = ["libraryname"]
     outputName = "Tests"
-    srcDirs = ["tests"]
+    sourceFiles = ["tests/*.cpp"]
     includePaths = ["include"]
     # The libraryPaths and libraries fields can be used to specify additional
     # libraries and paths to the build. This allows for linking against third 
@@ -33,5 +33,5 @@ flags = [
     buildRule = "exe"
     requires = ["libraryname"]
     outputName = "Applicatoin"
-    srcDirs = ["src"]
+    sourceFiles = ["src/*.cpp"]
     includePaths = ["include"]
